pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "fastapi>=0.116.1",
    "google-genai>=1.30.0",
    "openai>=1.99.9",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "supabase>=2.18.1",
    "uvicorn>=0.35.0",
]
